requirements.txt
# Runtime dependency management is defined in pyproject.toml.
# This file is optional and is not required for installation.
python-dotenv
requests
sqlalchemy
pydantic
alembic
psycopg2-binary
fastapi
uvicorn
cryptography
paramiko
